+++ BackendFinal/Areas/AdminArea/Controllers/BioController.cs
@@ -7,6 +7,7 @@ using Microsoft.AspNetCore.Mvc;
 
 namespace BackendFinal.Areas.AdminArea.Controllers
 {
+    [Area("AdminArea")]
     public class BioController : Controller
     {
         private readonly AppDbContext _appDbContext;
@@ -23,11 +24,13 @@ namespace BackendFinal.Areas.AdminArea.Controllers
             var bio = _appDbContext.Bios.FirstOrDefault();
             BioVM bioVM = new BioVM()
             {
-                //Photo
+                ImgUrl = bio.LogoUrl,
                 Address = bio.Address,
                 Email = bio.Email,
                 PhoneNumber = bio.PhoneNumber,
-                WorkingTime = bio.WorkingTime
+                WorkingTime = bio.WorkingTime,
+                GeneralInformation = bio.GeneralInformation,
+                AboutImgUrl = bio.AboutImgUrl
             };
             return View(bioVM);
         }
@@ -40,14 +43,25 @@ namespace BackendFinal.Areas.AdminArea.Controllers
             var existbio = _appDbContext.Bios.FirstOrDefault();
             if (existbio != null)
             {
-                string path = Path.Combine(_webHostEnvironment.WebRootPath, "img", existbio.LogoUrl);
-                DeleteHelper.DeleteFile(path);
-                existbio.LogoUrl = bioVM.Photo.FileName;
-                existbio.Email = bioVM.Email;
-                existbio.Address = bioVM.Address;
-                existbio.PhoneNumber = bioVM.PhoneNumber;
-                existbio.WorkingTime = bioVM.WorkingTime;
+                if (bioVM.Photo != null)
+                {
+                    string path = Path.Combine(_webHostEnvironment.WebRootPath, "img", existbio.LogoUrl);
+                    DeleteHelper.DeleteFile(path);
+                    existbio.LogoUrl = bioVM.Photo.FileName;
+                }
+                if (bioVM.AboutPhoto != null)
+                {
+                    string path = Path.Combine(_webHostEnvironment.WebRootPath, "img", existbio.AboutImgUrl);
+                    DeleteHelper.DeleteFile(path);
+                    existbio.AboutImgUrl = bioVM.AboutPhoto.FileName;
+                }
+                
             }
+            existbio.Email = bioVM.Email;
+            existbio.Address = bioVM.Address;
+            existbio.PhoneNumber = bioVM.PhoneNumber;
+            existbio.WorkingTime = bioVM.WorkingTime;
+            existbio.GeneralInformation = bioVM.GeneralInformation;
 
             _appDbContext.SaveChanges();
             return RedirectToAction("Index");
